changed data type to long

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,16 +7,16 @@
 
 int main(void)
 {
-	long long first, second, next, i;
+	long first, second, next, i;
 
 	first = 1;
 	second = 2;
 	next = first + second;
 
-	printf("%lld, %lld", first, second);
+	printf("%ld, %ld", first, second);
 	for (i = 3; i <= 50; i++)
 	{
-		printf(", %lld", next);
+		printf(", %ld", next);
 		first = second;
 		second = next;
 		next = first + second;
